Add request to bank client, link it up to ProcessPayment method

+++ src/PaymentGateway.Api/Models/Responses/PaymentDetailsResponse.cs
@@ -1,5 +1,3 @@
-using System.Numerics;
-
 namespace PaymentGateway.Api.Models.Responses;
 
 public abstract class PaymentDetailsResponse
@@ -8,5 +6,5 @@ public abstract class PaymentDetailsResponse
     public required string Status { get; init; }
     public required Card CardDetails { get; init; }
     public required string ISOCurrencyCode { get; init; }
-    public required BigInteger Amount { get; init; }
+    public required decimal Amount { get; init; }
 }
